Adding unit tests for MovieDetails feature

+++ data/src/main/java/com/example/data/moviedetails/source/local/MovieDetailsLocalDataSource.kt
@@ -9,7 +9,7 @@ class MovieDetailsLocalDataSource @Inject constructor(
     private val movieDetailsDao: MovieDetailsDao
 ) {
 
-    suspend fun getMovie(id: Int): MovieDetailsLocal? {
+    suspend fun getMovie(id: Long): MovieDetailsLocal? {
         return  try {
             movieDetailsDao.getMovie(id)
         }catch (error: IOException){
